refactor(SynchroContext): Moved to a dedicated file
